function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
autopep8==1.5.7
azure-common==1.1.27
azure-core==1.17.0
azure-functions==1.7.0
azure-servicebus==7.3.2
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-http-client==3.3.2
requests==2.26.0
requests-oauthlib==1.3.0
sendgrid==6.3.0
six==1.16.0
toml==0.10.2
uamqp==1.4.1
urllib3==1.26.6
